Move RTCMultiConnection logic out of Receiver into isolated Vue Component
